requirements.txt
groq
pytubefix
moviepy==1.0.3
SpeechRecognition
python-dotenv==0.19.0
numpy
requests
Pillow
decorator>=4.0.2
proglog
tqdm
flask==2.0.1
google-api-python-client==2.15.0
youtube_transcript_api==0.4.4
fastapi==0.68.0
uvicorn==0.15.0
jinja2==3.0.1
aiofiles==0.7.0
python-multipart==0.0.5 